+++ app/controllers/frontend/nimble/spells_controller.rb
@@ -7,7 +7,7 @@ module Frontend
 
       def index
         serialize_relation_v2(
-          ::Nimble::Feat.where(origin: 4),
+          ::Nimble::Feat.where(origin: [3, 4]),
           ::Nimble::SpellSerializer,
           :spells,
           cache_options: { key: "nimble_spells/#{I18n.locale}/0.5.3", expires_in: 24.hours }