+++ app/serializers/nimble/character_serializer.rb
@@ -8,13 +8,14 @@ module Nimble
                :skill_points, :modified_abilities, :initiative, :size, :abilities, :armor, :speed, :health, :wounds_max,
                :wounds_spent, :languages, :attacks, :features, :hit_die_max, :hit_die_spent, :hit_die, :saves, :subclass,
                :conditions, :separate_shield, :shield, :key_points, :secondary_points, :keys, :schools, :mana_max, :mana_spent,
-               :spell_level, :spells
+               :spell_level, :spells, :utility_spells_limit, :learned_spells, :save_dc
 
     delegate :skills, :modified_abilities, :initiative, :armor, :shield, :speed, :wounds_max, :attacks, :features, :hit_die_max,
-             :hit_die_spent, :hit_die, :saves, :keys, :schools, :mana_max, :spell_level, :spells, to: :decorator
+             :hit_die_spent, :hit_die, :saves, :keys, :schools, :mana_max, :spell_level, :spells, :utility_spells_limit, :save_dc,
+             to: :decorator
     delegate :data, to: :object
     delegate :level, :guide_step, :ancestry, :main_class, :skill_points, :abilities, :health, :wounds_spent, :languages, :size,
-             :subclass, :conditions, :separate_shield, :key_points, :secondary_points, :mana_spent, to: :data
+             :subclass, :conditions, :separate_shield, :key_points, :secondary_points, :mana_spent, :learned_spells, to: :data
 
     def provider
       'nimble'