+++ db/seeds_prod.rb
@@ -77,6 +77,7 @@ end
 [
   'feats/ancestries.json',
   'feats/berserker.json', 'feats/cheat.json', 'feats/commander.json', 'feats/hunter.json', 'feats/zephyr.json',
+  'feats/mage.json',
   'spells.json'
 ].each do |filename|
   JSON.parse(File.read("db/data_prod/nimble/#{filename}")).each do |item|